Move glyphslib module to __init__
